update chronicle backend
Currently we use the process start time as the events timestamp. From feedback this should be instead the UTCTimestamp as often times a process may be running for a while before a detection is emitted. This will help to remove the delta between the detection being emitted and the time that Chronicle logs the event as occurring.