feat: add eventType filter
* Fix issue with improper type checking

RELEVANT_EVENT_TYPES is a list, therefore the check to see if
'ALL' was in there was orginally did not work since it was
treating it as a string. We now check to see if the string 'ALL'
is in the list.

* Introduce streaming url eventType filter.

This change takes the set of RELEVANT_EVENT_TYPES by a backend
and uses that as a filter to the streaming url. This prevents
the need to have the FIG process ALL events, instead allows it
to process only the relevant events, reducing overhead.

* Minor grammar update