chore: enhance logic around using cred store
This commit changes the logic around how to use a credential store via
the configuration file. It's broken out into it's own section now, this
gives the user greater flexibility to be able to use a credential store
with different backends.

Also added some additional error handling in the credstore for
secrets_manager.