fix: make pip package actually work
This PR introduces the ability to install falcon-integration-gateway
from the pypi repository and actually use it. A couple key changes were
needed for this to work:
1. Needed to update the setup.py file to include the relevant
directories that make the application work (ie fig/ config/)
2. Updated the config class to check the package level for the
defaults.ini file as the fig needs this to instantiate
3. Updated the config class to allow using local config files for
overrides. Now you can pip install falcon-integration-gateway, and add
your config.ini file and let it rip.
4. Updated setup.py to reflect existing dependency with tlssyslog
package. It's been deprecated and won't work with python 3.12+.