fix: fixes issue with loading config when not using package
There was an issue with the latest changes to the config/__init__.py
files that were made to support the updated python package. The problem
was that on top of some errors with the python packaging config, we also
stopped checking for what happens when you are running this from either
the existing container or from the repository itself (aka, not using the
pip package).