fix: fixed issue with severity values from detections being incorrect
This was from the recent changes from DSE to EPPDSE event types. This
fix maps based on the severity name instead of value. This currently
maintains backwards compat for DSE events in some of the other backends
and is just a simpler way to deal with this. Also added better logging
to why an event is skipped. Unfortunately, since we support more than
just Detections, I don't want to add a bunch of fields that are only
relevant to one event type. Instead, we'll need to think through how
best to provide someone with additional information..perhaps a verbosity
level - 1 is what we have and 2 will print out the event payload itself?